third_party/blink/web_tests/external/wpt/content-security-policy/navigate-to/spv-only-sent-to-initiator.sub-expected.txt

This is a testharness.js-based test.
FAIL Test that no spv event is raised assert_unreached: Should not have raised a spv event Reached unreachable code
FAIL Violation report status OK. assert_equals: CSP report sent, but not expecting one. expected (undefined) undefined but got (object) object "[object Object]"
Harness: the test ran to completion.
